requirements.txt
fastapi==0.111.1
langchain==0.2.10
langchain-community==0.2.9
langchain-core==0.2.22
langchain-openai==0.1.17
langchain-text-splitters==0.2.2
langserve==0.2.2
openai==1.59.8
python-dotenv
sse_starlette==2.1.2